+++ NEWS.md
@@ -4,6 +4,7 @@
 * add GPL 3 License
 * add DOI to Readme
 * add argumwent *beta* to function updatePlantumlJar() to enable download from beta versions
+* added check if plantuml.jar has been downloaded - will download it automatically when not downloaded
 
 # plantuml 0.2.3
 * quote names of objects to avoid errors with non-text-characters